viral_bot/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viral-bot"
version = "1.0.0"
description = "Viral Health Post Idea Bot for aging.ai and avoidaging Instagram pages"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "openai>=1.6.0",
    "google-auth>=2.25.0",
    "google-api-python-client>=2.111.0",
    "gspread>=5.12.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "numpy>=1.26.0",
    "tenacity>=8.2.3",
    "apscheduler>=3.10.4",
    "fastapi>=0.108.0",
    "uvicorn>=0.25.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "python-dateutil>=2.8.2",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.9"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.8.0",
]

[project.scripts]
viral-bot = "viral_bot.__main__:cli"

[project.urls]
Homepage = "https://github.com/yourusername/viral-bot"
Documentation = "https://github.com/yourusername/viral-bot#readme"
Repository = "https://github.com/yourusername/viral-bot.git"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
